revisi nomor sisipan

+++ app/Http/Controllers/MeterAirController.php
@@ -32,6 +32,7 @@ class MeterAirController extends Controller
             'stand_kini' => 'required|numeric|min:0',
             'tagihan_bulan_lalu' => 'nullable|numeric'
         ]);
+
         $exists = MeterAir::where('user_id', $request->user_id)
             ->where('bulan', $request->bulan)
             ->where('tahun', $request->tahun)
@@ -40,24 +41,16 @@ class MeterAirController extends Controller
         if ($exists) {
             return back()
                 ->withInput()
-                ->with('error', '❌ Data meter air untuk warga ini pada bulan & tahun tersebut sudah ada.');
+                ->with('error', 'Data meter air untuk warga ini pada bulan & tahun tersebut sudah ada.');
         }
 
-        // hitung periode sebelumnya
-        $bulanPrev = $request->bulan - 1;
-        $tahunPrev = $request->tahun;
+        $prev = MeterAir::prevRecord(
+            $request->user_id,
+            $request->bulan,
+            $request->tahun
+        );
 
-        if ($bulanPrev == 0) {
-            $bulanPrev = 12;
-            $tahunPrev--;
-        }
-
-        $prev = MeterAir::where('user_id', $request->user_id)
-            ->where('bulan', $bulanPrev)
-            ->where('tahun', $tahunPrev)
-            ->first();
-
-        $standLama = $prev ? $prev->stand_kini : 0;
+        $standLama = $prev?->stand_kini ?? 0;
         $pemakaian = $request->stand_kini - $standLama;
 
         if ($pemakaian < 0) {
@@ -65,12 +58,11 @@ class MeterAirController extends Controller
                 'stand_kini' => 'Stand kini tidak boleh lebih kecil dari stand lama'
             ]);
         }
+        $hutang = $request->tagihan_bulan_lalu ?? 0;
 
-        // $totalBayar =
-        //     self::ABONEMEN +
-        //     ($pemakaian * self::TARIF_PER_M3) -
-        //     ($request->tagihan_bulan_lalu ?? 0);
-
+        if ($hutang > 0) {
+            $hutang = -1 * $hutang;
+        }
         MeterAir::create([
             'user_id' => $request->user_id,
             'bulan' => $request->bulan,
@@ -78,13 +70,13 @@ class MeterAirController extends Controller
             'stand_lama' => $standLama,
             'stand_kini' => $request->stand_kini,
             'pemakaian' => $pemakaian,
-            'tagihan_bulan_lalu' => $request->tagihan_bulan_lalu ?? 0,
-            // 'total_bayar' => $totalBayar,
+            'tagihan_bulan_lalu' => $hutang,
         ]);
 
         return redirect()->route('air.index')
-            ->with('success', '✅ Data meter berhasil disimpan');
+            ->with('success', 'Data meter berhasil disimpan');
     }
+
 
 
 
@@ -94,7 +86,7 @@ class MeterAirController extends Controller
         $tahun = $request->tahun ?? date('Y');
         $search = $request->search;
 
-        $query = MeterAir::with('user')
+        $query = MeterAir::with('user.warga')
             ->where('bulan', $bulan)
             ->where('tahun', $tahun);
 
@@ -164,16 +156,21 @@ class MeterAirController extends Controller
         //     self::ABONEMEN +
         //     ($pemakaian * self::TARIF_PER_M3) -
         //     ($request->tagihan_bulan_lalu ?? 0);
-
+        if ($meter->status_lunas) {
+            abort(403, 'Tagihan sudah lunas — tidak boleh diubah');
+        }
+        $hutang = $request->tagihan_bulan_lalu ?? 0;
+        if ($hutang > 0) {
+            $hutang = -1 * $hutang;
+        }
         $meter->update([
             'bulan' => $request->bulan,
             'tahun' => $request->tahun,
             'stand_kini' => $request->stand_kini,
             'pemakaian' => $pemakaian,
-            'tagihan_bulan_lalu' => $request->tagihan_bulan_lalu,
+            'tagihan_bulan_lalu' => $hutang,
             // 'total_bayar' => $totalBayar
         ]);
-
         return redirect()->route('air.index')
             ->with('success','Data meter berhasil diupdate');
     }
@@ -181,30 +178,49 @@ class MeterAirController extends Controller
 
     public function destroy(MeterAir $meter)
     {
+        if ($meter->status_lunas) {
+            abort(403, 'Data lunas tidak boleh dihapus');
+        }
+
         $meter->delete();
-        return redirect()->route('air.index')->with('success','Data meter berhasil dihapus');
+        return back()->with('success','Data meter berhasil dihapus');
     }
     public function getPrevStand(Request $request)
     {
-        $bulan = $request->bulan;
-        $tahun = $request->tahun;
+        $prev = MeterAir::prevRecord(
+            $request->user_id,
+            $request->bulan,
+            $request->tahun
+        );
 
-        $bulanPrev = $bulan - 1;
-        $tahunPrev = $tahun;
-
-        if ($bulanPrev == 0) {
-            $bulanPrev = 12;
-            $tahunPrev--;
+        if (!$prev) {
+            return response()->json([
+                'stand_lama' => 0,
+                'tagihan' => 0,
+                'status_lunas' => 1
+            ]);
         }
 
-        $prev = MeterAir::where('user_id', $request->user_id)
-            ->where('bulan', $bulanPrev)
-            ->where('tahun', $tahunPrev)
-            ->first();
+        $total = self::ABONEMEN +
+                ($prev->pemakaian * self::TARIF_PER_M3) -
+                ($prev->tagihan_bulan_lalu ?? 0);
 
         return response()->json([
-            'stand_lama' => $prev->stand_kini ?? 0
+            'stand_lama' => $prev->stand_kini,
+            'tagihan' => $prev->status_lunas ? 0 : $total,
+            'status_lunas' => $prev->status_lunas ? 1 : 0
         ]);
     }
 
+    public function setLunas(Request $request, $id)
+    {
+        $data = MeterAir::findOrFail($id);
+        $data->status_lunas = true;
+        $data->save();
+
+        return redirect()->route('air.index', [
+            'bulan' => $request->bulan,
+            'tahun' => $request->tahun,
+        ])->with('success', 'Tagihan ditandai LUNAS');
+    }
 }
